// src/app/api/v1/model/apiV1DevicesGet200ResponseInner.ts
/**
 * Rea Prime API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApiV1DevicesGet200ResponseInner { 
    name?: string;
    id?: string;
    state?: ApiV1DevicesGet200ResponseInner.StateEnum;
}
export namespace ApiV1DevicesGet200ResponseInner {
    export type StateEnum = 'connected' | 'disconnected';
    export const StateEnum = {
        Connected: 'connected' as StateEnum,
        Disconnected: 'disconnected' as StateEnum
    };
}